Refactor and clean up methods
