+++ test/test_helper.rb
@@ -11,4 +11,3 @@ require './lib/teams'
 require './lib/games_repo'
 require './lib/game_teams_repo'
 require './lib/teams_repo'
-require './test/game_repo_test'
